+++ src/commands/deploy.ts
@@ -84,7 +84,9 @@ export const command = new Command("deploy")
     'only deploy to specified, comma-separated targets (e.g. "hosting,storage"). For functions, ' +
       'can specify filters with colons to scope function deploys to only those functions (e.g. "--only functions:func1,functions:func2"). ' +
       "When filtering based on export groups (the exported module object keys), use dots to specify group names " +
-      '(e.g. "--only functions:group1.subgroup1,functions:group2)"' +
+      '(e.g. "--only functions:group1.subgroup1,functions:group2"). ' +
+      "When filtering based on codebases, use colons to specify codebase names " +
+      '(e.g. "--only functions:codebase1:func1,functions:codebase2:group1.subgroup1"). ' +
       "For data connect, can specify filters with colons to deploy only a service, connector, or schema" +
       '(e.g. "--only dataconnect:serviceId,dataconnect:serviceId:connectorId,dataconnect:serviceId:schema"). ',
   )
@@ -113,7 +115,13 @@ export const command = new Command("deploy")
       try {
         await requireHostingSite(options);
       } catch (err: unknown) {
-        if (err === errNoDefaultSite) {
+        const isPermissionError =
+          err instanceof FirebaseError &&
+          err.original instanceof FirebaseError &&
+          err.original.status === 403;
+        if (isPermissionError) {
+          throw err;
+        } else if (err === errNoDefaultSite) {
           createSite = true;
         }
       }