+++ src/commands/firestore-backups-schedules-update.ts
@@ -9,20 +9,19 @@ import { Emulators } from "../emulator/types";
 import { warnEmulatorNotSupported } from "../emulator/commandUtils";
 import { FirestoreOptions } from "../firestore/options";
 import { PrettyPrint } from "../firestore/pretty-print";
+import { FirebaseError } from "../error";
 
 export const command = new Command("firestore:backups:schedules:update <backupSchedule>")
   .description("Update a backup schedule under your Cloud Firestore database.")
-  .option("-rt, --retention <duration>", "duration string (e.g. 12h or 30d) for backup retention")
+  .option("--retention <duration>", "duration string (e.g. 12h or 30d) for backup retention")
   .before(requirePermissions, ["datastore.backupSchedules.update"])
   .before(warnEmulatorNotSupported, Emulators.FIRESTORE)
   .action(async (backupScheduleName: string, options: FirestoreOptions) => {
     const printer = new PrettyPrint();
+    const helpCommandText = "See firebase firestore:backups:schedules:update --help for more info.";
 
     if (!options.retention) {
-      logger.error(
-        "Missing required flag --retention. See firebase firestore:backups:schedules:update --help for more info",
-      );
-      return;
+      throw new FirebaseError(`Missing required flag --retention. ${helpCommandText}`);
     }
     const retention = calculateRetention(options.retention);