+++ src/dataconnect/schemaMigration.ts
@@ -12,6 +12,7 @@ import { Options } from "../options";
 import { FirebaseError } from "../error";
 import { needProjectId } from "../projectUtils";
 import { logLabeledBullet, logLabeledWarning, logLabeledSuccess } from "../utils";
+import * as experiments from "../experiments";
 import * as errors from "./errors";
 
 export async function diffSchema(schema: Schema): Promise<Diff[]> {
@@ -22,6 +23,8 @@ export async function diffSchema(schema: Schema): Promise<Diff[]> {
     databaseId,
     /* linkIfNotConnected=*/ false,
   );
+
+  setCompatibleMode(schema, databaseId, instanceName);
   try {
     await upsertSchema(schema, /** validateOnly=*/ true);
     logLabeledSuccess("dataconnect", `Database schema is up to date.`);
@@ -63,6 +66,9 @@ export async function migrateSchema(args: {
     databaseId,
     /* linkIfNotConnected=*/ true,
   );
+
+  setCompatibleMode(schema, databaseId, instanceName);
+
   try {
     await upsertSchema(schema, validateOnly);
     logger.debug(`Database schema was up to date for ${instanceId}:${databaseId}`);
@@ -113,6 +119,24 @@ export async function migrateSchema(args: {
     return diffs;
   }
   return [];
+}
+
+function setCompatibleMode(schema: Schema, databaseId: string, instanceName: string) {
+  if (experiments.isEnabled("fdccompatiblemode")) {
+    if (schema.primaryDatasource.postgresql?.schemaValidation) {
+      schema.primaryDatasource.postgresql.schemaValidation = "COMPATIBLE";
+    } else {
+      schema.primaryDatasource = {
+        postgresql: {
+          database: databaseId,
+          cloudSql: {
+            instance: instanceName,
+          },
+          schemaValidation: "COMPATIBLE",
+        },
+      };
+    }
+  }
 }
 
 function getIdentifiers(schema: Schema): {