+++ src/emulator/functionsEmulator.ts
@@ -540,12 +540,14 @@ export class FunctionsEmulator implements EmulatorInstance {
       };
       const userEnvs = functionsEnv.loadUserEnvs(userEnvOpt);
       const discoveredBuild = await runtimeDelegate.discoverBuild(runtimeConfig, environment);
-      const resolution = await resolveBackend(
-        discoveredBuild,
-        JSON.parse(firebaseConfig),
+      const resolution = await resolveBackend({
+        build: discoveredBuild,
+        firebaseConfig: JSON.parse(firebaseConfig),
         userEnvOpt,
         userEnvs,
-      );
+        nonInteractive: false,
+        isEmulator: true,
+      });
       const discoveredBackend = resolution.backend;
       const endpoints = backend.allEndpoints(discoveredBackend);
       prepareEndpoints(endpoints);
@@ -588,6 +590,20 @@ export class FunctionsEmulator implements EmulatorInstance {
     } else {
       this.workerPools[emulatableBackend.codebase].refresh();
     }
+
+    // Remove any old trigger definitions
+    const toRemove = Object.keys(this.triggers).filter((recordKey) => {
+      const record = this.getTriggerRecordByKey(recordKey);
+      if (force) {
+        return true; // We are going to load all of the triggers anyway, so we can remove everything
+      }
+      return !triggerDefinitions.some(
+        (def) =>
+          record.def.entryPoint === def.entryPoint &&
+          JSON.stringify(record.def.eventTrigger) === JSON.stringify(def.eventTrigger),
+      );
+    });
+    await this.removeTriggers(toRemove);
 
     // When force is true we set up all triggers, otherwise we only set up
     // triggers which have a unique function name
@@ -684,6 +700,13 @@ export class FunctionsEmulator implements EmulatorInstance {
           case Constants.SERVICE_STORAGE:
             added = this.addStorageTrigger(this.args.projectId, key, definition.eventTrigger);
             break;
+          case Constants.SERVICE_FIREALERTS:
+            added = await this.addFirealertsTrigger(
+              this.args.projectId,
+              key,
+              definition.eventTrigger,
+            );
+            break;
           default:
             this.logger.log("DEBUG", `Unsupported trigger: ${JSON.stringify(definition)}`);
             break;
@@ -750,6 +773,36 @@ export class FunctionsEmulator implements EmulatorInstance {
     }
   }
 
+  // Currently only cleans up eventarc and firealerts triggers
+  async removeTriggers(toRemove: string[]) {
+    for (const triggerKey of toRemove) {
+      const definition = this.triggers[triggerKey].def;
+      const service = getFunctionService(definition);
+      const key = this.getTriggerKey(definition);
+
+      switch (service) {
+        case Constants.SERVICE_EVENTARC:
+          await this.removeEventarcTrigger(
+            this.args.projectId,
+            key,
+            definition.eventTrigger as EventTrigger,
+          );
+          delete this.triggers[key];
+          break;
+        case Constants.SERVICE_FIREALERTS:
+          await this.removeFirealertsTrigger(
+            this.args.projectId,
+            key,
+            definition.eventTrigger as EventTrigger,
+          );
+          delete this.triggers[key];
+          break;
+        default:
+          break;
+      }
+    }
+  }
+
   addEventarcTrigger(projectId: string, key: string, eventTrigger: EventTrigger): Promise<boolean> {
     if (!EmulatorRegistry.isRunning(Emulators.EVENTARC)) {
       return Promise.resolve(false);
@@ -766,6 +819,78 @@ export class FunctionsEmulator implements EmulatorInstance {
       .then(() => true)
       .catch((err) => {
         this.logger.log("WARN", "Error adding Eventarc function: " + err);
+        return false;
+      });
+  }
+
+  removeEventarcTrigger(
+    projectId: string,
+    key: string,
+    eventTrigger: EventTrigger,
+  ): Promise<boolean> {
+    if (!EmulatorRegistry.isRunning(Emulators.EVENTARC)) {
+      return Promise.resolve(false);
+    }
+    const bundle = {
+      eventTrigger: {
+        ...eventTrigger,
+        service: "eventarc.googleapis.com",
+      },
+    };
+    logger.debug(`removeEventarcTrigger`, JSON.stringify(bundle));
+    return EmulatorRegistry.client(Emulators.EVENTARC)
+      .post(`/emulator/v1/remove/projects/${projectId}/triggers/${key}`, bundle)
+      .then(() => true)
+      .catch((err) => {
+        this.logger.log("WARN", "Error removing Eventarc function: " + err);
+        return false;
+      });
+  }
+
+  addFirealertsTrigger(
+    projectId: string,
+    key: string,
+    eventTrigger: EventTrigger,
+  ): Promise<boolean> {
+    if (!EmulatorRegistry.isRunning(Emulators.EVENTARC)) {
+      return Promise.resolve(false);
+    }
+    const bundle = {
+      eventTrigger: {
+        ...eventTrigger,
+        service: "firebasealerts.googleapis.com",
+      },
+    };
+    logger.debug(`addFirealertsTrigger`, JSON.stringify(bundle));
+    return EmulatorRegistry.client(Emulators.EVENTARC)
+      .post(`/emulator/v1/projects/${projectId}/triggers/${key}`, bundle)
+      .then(() => true)
+      .catch((err) => {
+        this.logger.log("WARN", "Error adding FireAlerts function: " + err);
+        return false;
+      });
+  }
+
+  removeFirealertsTrigger(
+    projectId: string,
+    key: string,
+    eventTrigger: EventTrigger,
+  ): Promise<boolean> {
+    if (!EmulatorRegistry.isRunning(Emulators.EVENTARC)) {
+      return Promise.resolve(false);
+    }
+    const bundle = {
+      eventTrigger: {
+        ...eventTrigger,
+        service: "firebasealerts.googleapis.com",
+      },
+    };
+    logger.debug(`removeFirealertsTrigger`, JSON.stringify(bundle));
+    return EmulatorRegistry.client(Emulators.EVENTARC)
+      .post(`/emulator/v1/remove/projects/${projectId}/triggers/${key}`, bundle)
+      .then(() => true)
+      .catch((err) => {
+        this.logger.log("WARN", "Error removing FireAlerts function: " + err);
         return false;
       });
   }