+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Pass down `functions.ignore` values in `firebase.json` to functions emulator so that supposedly ignored directories/files will not trigger reload. (#7414)
-- Fixes bug where secret values provided in env files were parsed as list (#7422)
+- Fixed an issue where functions deployment would fail if `firebase.json#extensions` was undefined. (#7575)